Establish convention for event handlers.
Using the top-level "else" to return Super was problematic for state handlers
that had transition guards, because it would not return the event to the super
state when the event did not enter the transition guard.

The solution is to establish a new convention: all events that are handled will
return from the state handler function right from the state handler, and the
"return Super" will be unguarded at the end of every state handler function.
This way, when an event has an unfulfilled transition guard, we will still
return the event to the super state for handling.