Introduce ao class.
The ao class implements the basic features of an active object. It contains an
event queue and an event post interface with the post() method. It takes an
io_loop as an argument, which is used to insert itself on to a central io loop.

When the application posts events to the ao, the event is put on the event
queue. In the next iteration through the event loop, the event is popped off
the queue and dispatched to the ao's hsm, where the event is handled. When the
event is finished handling, it returns thread control to the io loop.

For now, we couple ao to the tornado io loop. Later we can add a generic io
loop module to the ao package and support different types of async io loops,
such as tornado, asyncio, twisted, etc.