Make event class more flexible for simple events.
Events will always be defined by their sig property, but we don't need to
require application developers to subclass event for every type of event they
want. For simple events, that do not require any data or extra arguments, it is
easier to simply use the base event class with a simple string or other value
for the sig property, instead of the event class type. If developers want to
create more complicated events, they still can, and the sig property will
default to being the event class type.